Reworked gravity
Now every object is stored in a list. Every object in range of 10 attracts. Trying to implement creation of objects at runtime.

+++ Assets/code/PlanetObjectCode.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class planetObject : MonoBehaviour
 {
     public string planetName;
-    public int density = 1;
+    public float density = 1;
